Move "gpt4o.typing" to "lib".
