+++ pylib-v1/ailets/models/gpt4o/lib/typing.py
